+++ src/Extensions/OptionZipArchiveExtensions.cs
@@ -1,6 +1,6 @@
 using System.IO.Compression;
 
-namespace Ametrin.Optional.Extensions;
+namespace Ametrin.Optional;
 
 public static class OptionZipArchiveExtensions
 {