+++ src/Operations/Consume.cs
@@ -1,3 +1,6 @@
+using System.Diagnostics;
+using System.Diagnostics.CodeAnalysis;
+
 namespace Ametrin.Optional;
 
 partial struct Option
@@ -215,5 +218,119 @@ partial struct RefOption<TValue>
         }
 
         return _hasValue;
+    }
+}
+
+partial class OptionTupleExtensions
+{
+    public static Option Consume<T1, T2>(this (Option<T1>, Option<T2>) options, Action<T1, T2>? success = null, Action? error = null)
+    {
+        var (a, b) = options;
+        if (a._hasValue && b._hasValue)
+        {
+            success?.Invoke(a._value, b._value);
+            return true;
+        }
+        else
+        {
+            error?.Invoke();
+            return false;
+        }
+    }
+
+    public static Option Consume<T1, T2, TArg>(this (Option<T1>, Option<T2>) options, TArg arg, Action<T1, T2, TArg>? success = null, Action<TArg>? error = null)
+        where TArg : allows ref struct
+    {
+        var (a, b) = options;
+        if (a._hasValue && b._hasValue)
+        {
+            success?.Invoke(a._value, b._value, arg);
+            return true;
+        }
+        else
+        {
+            error?.Invoke(arg);
+            return false;
+        }
+    }
+
+    public static ErrorState Consume<T1, T2>(this (Result<T1>, Result<T2>) options, Action<T1, T2>? success = null, Action<Exception>? error = null)
+    {
+        var (a, b) = options;
+
+        if (a._hasValue && b._hasValue)
+        {
+            success?.Invoke(a._value, b._value);
+            return default;
+        }
+
+        if (!ErrorState.CombineErrors(a.ToErrorState(), b.ToErrorState()).Branch(out var err))
+        {
+            error?.Invoke(err);
+            return err;
+        }
+
+        throw new UnreachableException();
+    }
+
+    public static ErrorState Consume<T1, T2, TArg>(this (Result<T1>, Result<T2>) options, TArg arg, Action<T1, T2, TArg>? success = null, Action<Exception, TArg>? error = null)
+        where TArg : allows ref struct
+    {
+        var (a, b) = options;
+
+        if (a._hasValue && b._hasValue)
+        {
+            success?.Invoke(a._value, b._value, arg);
+            return default;
+        }
+
+        if (!Result.CombineErrors(a, b).Branch(out var err))
+        {
+            error?.Invoke(err, arg);
+            return err;
+        }
+
+        throw new UnreachableException();
+    }
+
+    [Experimental("AmOptional000")]
+    public static ErrorState<E> Consume<T1, T2, E>(this (Result<T1, E>, Result<T2, E>) options, Func<E, E, E> errorCombiner, Action<T1, T2>? success = null, Action<E>? error = null)
+    {
+        var (a, b) = options;
+
+        if (a._hasValue && b._hasValue)
+        {
+            success?.Invoke(a._value, b._value);
+            return default;
+        }
+
+        if (!Result.CombineErrors(a, b, errorCombiner).Branch(out var err))
+        {
+            error?.Invoke(err);
+            return err;
+        }
+
+        throw new UnreachableException();
+    }
+
+    [Experimental("AmOptional000")]
+    public static ErrorState<E> Consume<T1, T2, E, TArg>(this (Result<T1, E>, Result<T2, E>) options, TArg arg, Func<E, E, TArg, E> errorCombiner, Action<T1, T2, TArg>? success = null, Action<E, TArg>? error = null)
+        where TArg : allows ref struct
+    {
+        var (a, b) = options;
+
+        if (a._hasValue && b._hasValue)
+        {
+            success?.Invoke(a._value, b._value, arg);
+            return default;
+        }
+
+        if (!Result.CombineErrors(a, b, arg, errorCombiner).Branch(out var err))
+        {
+            error?.Invoke(err, arg);
+            return err;
+        }
+
+        throw new UnreachableException();
     }
 }