+++ src/Parsing/GenerateISpanParsableAttribute.cs
@@ -1,0 +1,4 @@
+namespace Ametrin.Optional.Parsing;
+
+[AttributeUsage(AttributeTargets.Class | AttributeTargets.Struct, AllowMultiple = false, Inherited = false)]
+public sealed class GenerateISpanParsableAttribute : Attribute;